requirements.txt
openai
langchain
langchain-openai
langchain-community
langchain-experimental
langchain_mistralai
langchain_cohere==0.1.5
langgraph
langserve[all]
langchain-cli
botbuilder-integration-aiohttp>=4.14.4
tiktoken
docx2txt
pillow
pypdf
tenacity
sqlalchemy
pyodbc
tabulate
azure-cosmos
streamlit
python-dotenv
azure-ai-formrecognizer
azure-storage-blob
beautifulsoup4
pydantic==1.10.13
fastapi